evaluation and themes modified

+++ server_api/src/modules/calendar/calendar.service.ts
@@ -1,3 +1,4 @@
+import { Call } from "../call/call.model";
 import { CalendarStatus } from "./calendar.enum";
 import { Calendar } from "./calendar.model";
 
@@ -32,6 +33,8 @@ export class CalendarService {
     static async deleteCalendar(id: string) {
         const calendar = await Calendar.findById(id);
         if (!calendar) throw new Error("Calendar not found");
+        const referencedByCall = await Call.exists({ calendar: calendar._id });
+        if (referencedByCall) throw new Error(`Can not delete ${calendar.year}, it is referenced in call.`);
         return await calendar.deleteOne();
     }
 }
